Add 'list' command base
