+++ src/store.js
@@ -10,9 +10,9 @@ const store = configureStore({
     terminal: terminalReducer,
     history: historyReducer,
   },
-  preloadedState: loadFromLocalStorage(),
+  // preloadedState: loadFromLocalStorage(),
 });
 
-store.subscribe(() => saveToLocalStorage(store.getState()));
+// store.subscribe(() => saveToLocalStorage(store.getState()));
 
 export default store;
